refactor: standardize logging with new Logger interface implementations
- Remove the old logger implementations for Slog and Zap from the logger.go file.
- Introduce the Logger interface to standardize logging methods.
- Add a new file for Slog logger implementation.
- Add a new file for Zap logger implementation.
